backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector search and retrieval
sentence-transformers==2.3.1
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.26.3

# Database ORM
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0

# BM25 retrieval
rank-bm25==0.2.2

# Cross-encoder for reranking
torch==2.1.2
transformers==4.36.2

# Environment management
python-dotenv==1.0.0

# Redis for real-time synchronization
redis==5.0.1
aioredis==2.0.1

# Temporal for workflow orchestration and locking
temporalio==1.5.1

# System monitoring
psutil==5.9.6
